+++ RestaurantProject/db/schema/imports/reset_tables.py
@@ -1,8 +1,12 @@
-# ensure all models are registered with Base.metadata
 from db.schema import *  # imports Account, Reservation, Events, etc.
 from db.server import Base, engine
 
-Base.metadata.drop_all(bind=engine)
-Base.metadata.create_all(bind=engine)
+def reset_all():
+    # ensure all models are registered with Base.metadata
 
-print("Tables reset!")
+    Base.metadata.drop_all(bind=engine)
+    Base.metadata.create_all(bind=engine)
+
+    print("Tables reset!")
+
+reset_all()