+++ changelog.md
@@ -1,3 +1,17 @@
+## TO-DO:
+* Fix Some Filters
+* Add Menu Items
+* Add Route Decorators That Specifically Make Pages Available Based On Role
+* Fix Routing + Merge ***User Land*** + ***Restaurant Listings*** 
+
+# CHANGES [v3.1.1]:
+* Changed `import.py` To Also Reset Tables Beforehand To Ensure
+    * Also Fixed `Reservation` Import csv
+* Fixed `get_distance_miles`
+* Fixed Cookies + User Routing
+* Started Migrating `/delete` Into The Profile Page
+* Updated `requirements.txt` 
+
 # CHANGES [v3.1.0]: 
 * Added Full Date and Time Field to Reservation
 * Cookie for Login and Signup to Stay Logged In
@@ -6,10 +20,6 @@
 * Started Filter Functionality
 * Reverted Restaurant Routes to Use The RID instead of Name
     * Makes Sure Duplicate Names Can Have Seperate Pages and Links
-## TO-DO:
-* Fix Some Filters
-* Add Menu Items
-* Add Route Decorators That Specifically Make Pages Available Based On Role
 
 # CHANGES [v2.1.0]: Templating and Page Generation
 
@@ -22,7 +32,7 @@
 # To-Do List:
 * Add More Client Side Locking
 * <i>Make It Look Cool</i>
-* Conformation for Delete
+* Confirmation for Delete
 # COMMIT 1
 ## **NEW TABLE**
     Testing with a Menu Table to load Menu Items Per Page